Fix level 4 errors

+++ wcfsetup/install/files/lib/form/RegisterActivationForm.class.php
@@ -110,7 +110,7 @@ final class RegisterActivationForm extends AbstractFormBuilderForm
      */
     public function show()
     {
-        if (!(REGISTER_ACTIVATION_METHOD & User::REGISTER_ACTIVATION_USER)) {
+        if (!((int)REGISTER_ACTIVATION_METHOD & User::REGISTER_ACTIVATION_USER)) {
             throw new IllegalLinkException();
         }
