+++ wcfsetup/install/files/lib/system/clipboard/action/UserClipboardAction.class.php
@@ -325,7 +325,7 @@ class UserClipboardAction extends AbstractClipboardAction
         // check permissions
         if (
             !WCF::getSession()->getPermission('admin.user.canEnableUser')
-            || !(REGISTER_ACTIVATION_METHOD & User::REGISTER_ACTIVATION_USER)
+            || !((int)REGISTER_ACTIVATION_METHOD & User::REGISTER_ACTIVATION_USER)
         ) {
             return [];
         }
